move TaskTracker to trace
